IN PROGRESS: added METS into new pattern
