+++ app/models/variations_document.rb
@@ -3,4 +3,6 @@ class VariationsDocument
   include PreingestableDocument
   LOCAL_RECORD_CLASS = ::IuMetadata::VariationsRecord
   SOURCE_TITLE = ['Variations XML']
+
+  delegate :default_attributes, to: :local_record
 end
